np.float32 => np.float64

+++ pyaf/TS/SignalDecomposition_AR.py
@@ -204,7 +204,7 @@ class cAutoRegressiveEstimator:
         logger = tsutil.get_pyaf_logger();
         self.mARFrame = pd.DataFrame();
         self.mTimeInfo.addVars(self.mARFrame);
-        self.mCycleFrame[cycle_residue] = self.mCycleFrame[cycle_residue].astype(np.float32)            
+        self.mCycleFrame[cycle_residue] = self.mCycleFrame[cycle_residue]            
         self.mARFrame[cycle_residue] = self.mCycleFrame[cycle_residue]            
 
         self.mDefaultValues = {};
